feat(safeguard): 引入冷却与尝试限制的自定义异常工厂机制
- 为 `AttemptLimit` 和 `Cooldown` 注解增加了自定义异常工厂支持，允许用户通过 `factory` 参数指定
异常创建逻辑。同时扩展了注解属性以支持更多业务场景配置，如 `scene`、`infoCode` 和`template` 等
- 新增相关上下文类 `AttemptLimitContext`与 `CooldownContext`用于传递
构建异常所需的上下文信息。
- 更新了切面逻辑以从 Spring 上下文中获取或实例化异常工厂，
- 修改 `Key.full()` 键格式分隔符为单冒号。- 调整 `AttemptCheck` 与 `AttemptLimitResult` 模型，增加剩余尝试次数与容量字段。
- 升级 `max` 参数类型由 Int 改为 Long以适应更大范围配置。
- 在 build.gradle.kts 中添加 spring-boot-starter-aop 依赖以支持切面功能。

+++ forgeboot-safeguard/forgeboot-safeguard-core/src/main/kotlin/com/gewuyou/forgeboot/safeguard/core/extensions/MapExtensions.kt
@@ -23,11 +23,11 @@ package com.gewuyou.forgeboot.safeguard.core.extensions
 import java.time.Duration
 
 /**
- * 将Map<Int, Duration>转换为CSV格式的字符串
+ * 将Map<Long, Duration>转换为CSV格式的字符串
  *
  * @return 返回格式为"k1:v1,k2:v2,..."的字符串，其中k是键，v是持续时间的毫秒值
  *         结果按键的升序排列
  */
-fun Map<Int, Duration>.toCsv(): String =
+fun Map<Long, Duration>.toCsv(): String =
     entries.sortedBy { it.key } // 按键排序
         .joinToString(",") { (k, v) -> "$k:${v.toMillis()}" } // 转换为key:millis格式并用逗号连接
